refactor(codegen): extract bytes lowering helpers
